Исправленные файлы заданий №2,5

+++ Practice/GarbisO/lec3 t2.py
@@ -1,4 +1,4 @@
-distance = float(input("Введите пройденное расстояние в метрах: "))
-time = float(input("Введите затраченное время в секундах: "))
-average_speed = float(distance/time)
+distance = int(input("Введите пройденное расстояние в метрах: "))
+time = int(input("Введите затраченное время в секундах: "))
+average_speed = distance/time
 print(f"Средняя скорость движения: {average_speed} м/с")
